FIX: Tweak to last commit to ensure alternative JPGs are not delete
git-svn-id: http://svn.resourcespace.com/svn/rs/trunk@9865 c08608d7-6e46-0410-86ca-f2a6f1370df5

+++ include/image_processing.php
@@ -2743,7 +2743,7 @@ function delete_previews($resource,$alternative=-1)
 	
 	$fullsizejpgpath=get_resource_path($resource,true,"",false,"jpg",-1,1,false,"",$alternative);			
 	# Delete the full size original if not a JPG resource
-	if($extension!="" && strtolower($extension)!="jpg" && file_exists($fullsizejpgpath))
+	if($extension!="" && strtolower($extension)!="jpg" && file_exists($fullsizejpgpath) && $alternative==-1)
 		{
 		unlink($fullsizejpgpath);
 		}
